+++ spotinst_sdk2/__init__.py
@@ -12,6 +12,7 @@ from spotinst_sdk2.clients.mrscaler import *
 from spotinst_sdk2.clients.subscription import *
 from spotinst_sdk2.clients.setup import *
 from spotinst_sdk2.clients.ocean_cd import *
+from spotinst_sdk2.clients.hpc import *
 
 class SpotinstSession:
     def __init__(self,
@@ -63,7 +64,9 @@ class SpotinstSession:
                                                              log_level=log_level, user_agent=user_agent,
                                                              timeout=timeout),
             "subscription": SubscriptionClient(session=self.session, print_output=print_output, log_level=log_level,
-                                               user_agent=user_agent, timeout=timeout)
+                                               user_agent=user_agent, timeout=timeout),
+            "hpc": HPCAwsClient(session=self.session, print_output=print_output,
+                                                    log_level=log_level, user_agent=user_agent, timeout=timeout)
         }
 
         return switcher.get(service, "Invalid Service")